feat: Create separate mobile page for blog posts
- Creates a new, simplified blog post page for mobile users to improve readability and performance on smaller devices.
- Moves the original, complex layout to a dedicated desktop component.
- Implements server-side user-agent detection in `page.js` to conditionally render either the mobile or desktop version of the page.
- Ensures feature parity by including JSON-LD structured data in both mobile and desktop versions for consistent SEO.
